Left navigation is now rendering the same as before
Migrating the sidebar from ForTrainingTemplate.php to SkinMustache: Now
it should render the same as before.
Portlet.mustache is essentially taking care of the same stuff as
renderPortal() was doing before.
Our previous hackish solution to enable links in navigation block
headings was not migrated. TODO make an issue out of this and bring back
that feature in a better way.