Restore right navigation
Extra mustache template as the HTML is a bit different than
Portlet.mustache.
Before there was an extra <span> around the <a> which had the
padding-top: 1.25em. However changing the HTML to include that <span>
again would require some investigation and hacks in PHP
because that's inside the data-portlets.data-namespaces.html-items (with
HTML code for the all list items together).
So I added the padding-top: 1.25em to div.vectorTabs li a
to have the same styling as before. I hope that doesn't have any side
effects...
Also removed some dead code.